+++ src/repositories/index.js
@@ -1,6 +1,6 @@
 module.exports = {  
     CampaignRepository:require('./campaign-repository'),
-    UserRepository: require('./user-repository'),
+    UserRepository: require('./users/user-repository'),
     ProductRepository: require('./product-repository'),
     LocationRepository : require('./location-repository'),
     DeviceRepository:require('./device-repository'),
