+++ src/services/users/user-service.js
@@ -1,4 +1,4 @@
-const UserRepository = require('../../repositories/user-repository');
+const UserRepository = require('../../repositories/users/user-repository');
 const { User } = require('../../models');
